Update umqtt
umqtt.simple ping function has been rewritten to guarantee either error if PINGRESP is not received.

umqtt.robust has been rewritten with a universal retry function based on an example from @pfalcon. It also requires now to specify the number of retries so it doesn't enter in an infinite retry loop.